refactor: move read-only cache interfaces to api package
