+++ operator-framework-core/src/main/java/io/javaoperatorsdk/operator/api/reconciler/IndexedResourceCache.java
@@ -1,0 +1,9 @@
+package io.javaoperatorsdk.operator.api.reconciler;
+
+import java.util.List;
+
+import io.fabric8.kubernetes.api.model.HasMetadata;
+
+public interface IndexedResourceCache<T extends HasMetadata> extends ResourceCache<T> {
+  List<T> byIndex(String indexName, String indexKey);
+}
