+++ operator-framework-core/src/main/java/io/javaoperatorsdk/operator/api/reconciler/ResourceCache.java
@@ -1,9 +1,10 @@
-package io.javaoperatorsdk.operator.processing.event.source;
+package io.javaoperatorsdk.operator.api.reconciler;
 
 import java.util.function.Predicate;
 import java.util.stream.Stream;
 
 import io.fabric8.kubernetes.api.model.HasMetadata;
+import io.javaoperatorsdk.operator.processing.event.source.Cache;
 
 @SuppressWarnings("unchecked")
 public interface ResourceCache<T extends HasMetadata> extends Cache<T> {
