--- operator-framework-core/src/main/java/io/javaoperatorsdk/operator/processing/event/source/IndexedResourceCache.java
@@ -1,9 +1,0 @@
-package io.javaoperatorsdk.operator.processing.event.source;
-
-import io.fabric8.kubernetes.api.model.HasMetadata;
-
-import java.util.List;
-
-public interface IndexedResourceCache<T extends HasMetadata> extends ResourceCache<T> {
-    List<T> byIndex(String indexName, String indexKey);
-}
