+++ operator-framework-core/src/main/java/io/javaoperatorsdk/operator/processing/event/source/IndexerResourceCache.java
@@ -5,6 +5,7 @@ import java.util.Map;
 import java.util.function.Function;
 
 import io.fabric8.kubernetes.api.model.HasMetadata;
+import io.javaoperatorsdk.operator.api.reconciler.IndexedResourceCache;
 
 public interface IndexerResourceCache<T extends HasMetadata> extends IndexedResourceCache<T> {
